- unit tests for LocalSquareBlockMinMax

+++ main/ip/src/boofcv/abst/filter/binary/LocalSquareBlockMinMaxBinaryFilter.java
@@ -37,10 +37,10 @@ public class LocalSquareBlockMinMaxBinaryFilter<T extends ImageSingleBand>
 	ThresholdSquareBlockMinMax alg;
 	ImageType<T> imageType;
 
-	public LocalSquareBlockMinMaxBinaryFilter(float textureThreshold, int requestedBlockWidth, double scale , boolean down, Class<T> imageType ) {
+	public LocalSquareBlockMinMaxBinaryFilter(double textureThreshold, int requestedBlockWidth, double scale , boolean down, Class<T> imageType ) {
 
 		if( imageType == ImageFloat32.class )
-			this.alg = new ThresholdSquareBlockMinMax_F32(textureThreshold,requestedBlockWidth,(float)scale,down);
+			this.alg = new ThresholdSquareBlockMinMax_F32((float)textureThreshold,requestedBlockWidth,(float)scale,down);
 		else if( imageType == ImageUInt8.class )
 			this.alg = new ThresholdSquareBlockMinMax_U8(textureThreshold,requestedBlockWidth,scale,down);
 		else
